Fix: Separado la creacion de datos por default de la data en la base de datos, ahora se separa a un archivo seed.js que se encarga de cargar todos estos datos, esto es para mas facil debugging, para ejecutarlo se usa npm run seed
