+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "start": "node index.js",
     "dev": "nodemon index.js",
-    "test": "mocha"
+    "test": "mocha",
+    "seed": "node ./src/seed.js"
   },
   "keywords": [],
   "author": "",
