+++ src/datos/roles.js
@@ -1,0 +1,26 @@
+const roles = [
+  {
+    id: 'a83af560-8a48-4edd-83eb-6bd3f7e97861',
+    name: 'sUser',
+  },
+  {
+    id: '633c2cda-d0f1-40f8-b07c-041a8cc217be',
+    name: 'Admin',
+  },
+  {
+    id: '606c0802-5332-4531-9189-eac84e6fcceb',
+    name: 'Profesor',
+  },
+  {
+    id: '5d3709ba-3a27-48cc-8a75-256338684cee',
+    name: 'Alumno',
+  },
+  {
+    id: '8e70a48b-4f1e-4fe9-93ed-5f41d48ee14d',
+    name: 'Tutor',
+  },
+]
+
+module.exports = {
+  roles,
+}
